requirements.txt
accelerate==0.21.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
blessed==1.21.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
einops==0.6.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gpustat==1.1.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
preshed==3.0.9
psutil==7.0.0
pycocotools==2.0.10
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.1.99
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
statsmodels==0.14.5
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==0.6.13
tokenizers==0.13.3
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.31.0
triton==3.3.0
typeguard==4.4.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
zipp==3.21.0
